feat(custom progress): refactor and clean code
