+++ src/shared/components/statistics/components/add-info/AddInfo.tsx
