+++ src/shared/components/statistics/components/add-info/addInfo.module.css
