+++ src/shared/components/statistics/components/error-progress-pagination/errorProgressPagination.module.css
