+++ src/shared/components/statistics/statistics.module.css
